requirements.txt
torch
torchvision
numpy
opencv-python
pillow
ffmpeg-python
tqdm
matplotlib
